Rename LSTM kernels to LSTM_INTERNAL  (#1699)
BUG=http://b/266447344

Rename LSTM kernels to *_LSTM_internal_* etc, and fix the resulting build errors. 

This is the first step to substitute the TFLite ported LSTM kernels with the template based new implementations.
