+++ tensorflow/lite/micro/kernels/BUILD
@@ -131,7 +131,7 @@ cc_library(
 cc_library(
     name = "lstm_eval_test_lib",
     hdrs = [
-        "lstm_eval_test.h",
+        "lstm_eval_internal_test.h",
     ],
     deps = [
         ":kernel_util",
@@ -228,7 +228,7 @@ tflm_kernel_cc_library(
         "logical_common.cc",
         "logistic.cc",
         "logistic_common.cc",
-        "lstm_eval.cc",
+        "lstm_eval_internal.cc",
         "maximum_minimum.cc",
         "mirror_pad.cc",
         "mul.cc",
@@ -268,7 +268,7 @@ tflm_kernel_cc_library(
         "tanh.cc",
         "transpose.cc",
         "transpose_conv.cc",
-        "unidirectional_sequence_lstm.cc",
+        "unidirectional_sequence_lstm_internal.cc",
         "unpack.cc",
         "var_handle.cc",
         "while.cc",
@@ -287,7 +287,7 @@ tflm_kernel_cc_library(
         "leaky_relu.h",
         "logical.h",
         "logistic.h",
-        "lstm_eval.h",
+        "lstm_eval_internal.h",
         "lstm_shared.h",
         "micro_ops.h",
         "mul.h",
@@ -356,9 +356,9 @@ tflm_kernel_cc_library(
 )
 
 cc_library(
-    name = "unidirectional_sequence_lstm_test_config",
-    srcs = ["unidirectional_sequence_lstm_test_config.cc"],
-    hdrs = ["unidirectional_sequence_lstm_test_config.h"],
+    name = "unidirectional_sequence_lstm_internal_test_config",
+    srcs = ["unidirectional_sequence_lstm_internal_test_config.cc"],
+    hdrs = ["unidirectional_sequence_lstm_internal_test_config.h"],
     deps = ["//tensorflow/lite/c:common"],
 )
 
@@ -915,9 +915,9 @@ cc_test(
 )
 
 cc_test(
-    name = "lstm_eval_test",
+    name = "lstm_eval_internal_test",
     srcs = [
-        "lstm_eval_test.cc",
+        "lstm_eval_internal_test.cc",
     ],
     tags = [
         "noasan",  # TODO(b/266028903): fix test with ubsan
@@ -1354,9 +1354,9 @@ cc_test(
 )
 
 cc_test(
-    name = "unidirectional_sequence_lstm_test",
+    name = "unidirectional_sequence_lstm_internal_test",
     srcs = [
-        "unidirectional_sequence_lstm_test.cc",
+        "unidirectional_sequence_lstm_internal_test.cc",
     ],
     deps = [
         ":kernel_runner",
@@ -1366,7 +1366,7 @@ cc_test(
         "//tensorflow/lite/micro:micro_utils",
         "//tensorflow/lite/micro:op_resolvers",
         "//tensorflow/lite/micro:test_helpers",
-        "//tensorflow/lite/micro/kernels:unidirectional_sequence_lstm_test_config",
+        "//tensorflow/lite/micro/kernels:unidirectional_sequence_lstm_internal_test_config",
         "//tensorflow/lite/micro/testing:micro_test",
     ],
 )
